test: add test cases
