--- packages/mp-loader/src/transpiler/__tests__/generate.js
@@ -1,10 +1,0 @@
-const generate = require('../generate');
-const modules = require('../transpileModules');
-
-const transpilerOptions = { modules };
-
-describe('Transpiler generate', () => {
-  it('should use a:for with other rules', () => {
-
-  });
-});
